+++ src/ldap/testing.rs
@@ -2,7 +2,6 @@ use super::*;
 use maplit::hashmap;
 
 struct ExpectedLdapPaths {
-    pub ldap_dc: String,
     pub ldap_bind: String,
     pub ldap_base: String,
 }
@@ -16,7 +15,6 @@ fn should_give_correct_ldap_paths() {
         let actual = LDAPConfig::new(&given, &Some("xxx".to_owned()), &Some("xxxx".to_owned()));
         assert_case(
             ExpectedLdapPaths {
-                ldap_dc: "".to_owned(),
                 ldap_base: "".to_owned(),
                 ldap_bind: "cn=xxx".to_owned(),
             },
@@ -37,7 +35,6 @@ fn should_give_correct_ldap_paths() {
 
         assert_case(
             ExpectedLdapPaths {
-                ldap_dc: "dc=example,dc=com".to_owned(),
                 ldap_base: "ou=it,ou=people,dc=example,dc=com".to_owned(),
                 ldap_bind: "uid=alice,ou=special,dc=example,dc=com".to_owned(),
             },
@@ -52,7 +49,6 @@ fn should_give_correct_ldap_paths() {
         let actual = LDAPConfig::new(&given, &Some("alice".to_owned()), &Some("xxxx".to_owned()));
         assert_case(
             ExpectedLdapPaths {
-                ldap_dc: "dc=example,dc=com".to_owned(),
                 ldap_base: "dc=example,dc=com".to_owned(),
                 ldap_bind: "cn=alice,dc=example,dc=com".to_owned(),
             },
@@ -68,7 +64,6 @@ fn should_give_correct_ldap_paths() {
         let actual = LDAPConfig::new(&given, &Some("alice".to_owned()), &Some("xxxx".to_owned()));
         assert_case(
             ExpectedLdapPaths {
-                ldap_dc: "dc=example,dc=com".to_owned(),
                 ldap_base: "aaa=department,bbb=level,dc=example,dc=com".to_owned(),
                 ldap_bind: "cn=alice,dc=example,dc=com".to_owned(),
             },
@@ -78,15 +73,13 @@ fn should_give_correct_ldap_paths() {
 
     fn assert_case(expected: ExpectedLdapPaths, actual: &LDAPConfig) {
         assert_eq!(
-            expected.ldap_dc, actual.ldap_dc,
-            "Did not produce correct ldap dc."
-        );
-        assert_eq!(
-            expected.ldap_base, actual.ldap_base,
+            expected.ldap_base,
+            actual.base(),
             "Did not produce correct base with  org path + dc."
         );
         assert_eq!(
-            expected.ldap_bind, actual.ldap_bind,
+            expected.ldap_bind,
+            actual.bind(),
             "Correct user binding, prefix_user + user + bind_org + dc"
         );
     }